server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
pandas==2.1.3
numpy==1.26.2
redis==5.0.1
rq==1.15.1
werkzeug==3.0.1
python-dotenv==1.0.0
openpyxl==3.1.2